+++ src_tests/common_code_for_tests.py
@@ -1,6 +1,6 @@
 __author__ = 'Tomasz Rybotycki'
 
-from numpy import ndarray, zeros
+from numpy import ndarray, zeros, int64
 
 from src.Boson_Sampling_Utilities import generate_possible_outputs
 from src.BosonSamplingSimulator import BosonSamplingSimulator
@@ -30,7 +30,7 @@ class ApproximateDistributionCalculator:
 
         simulator = BosonSamplingSimulator(self.strategy)
 
-        outcomes_probabilities = zeros(len(possible_outcomes))
+        outcomes_probabilities = zeros(len(possible_outcomes), dtype=int64)
 
         samples = simulator.get_classical_simulation_results(self.configuration.initial_state, samples_number)
         for sample in samples:
